Create ResourceFilterSearch
